added many-to-many association table, added override methods for data classes to prevent infinite loops for some methods

+++ src/main/kotlin/io/cyborgsquirrel/entity/LocationConfigEntity.kt
@@ -8,7 +8,7 @@ import io.micronaut.serde.annotation.Serdeable
 data class LocationConfigEntity (
     @Id
     @GeneratedValue
-    var id: Long = -1,
+    var id: Int = -1,
 
     @MappedProperty("lat")
     var latitude: String? = null,
